requirements.txt
apturl==0.5.2
attrs==19.3.0
autopep8==1.6.0
awscli==1.18.69
bcrypt==3.1.7
blinker==1.4
botocore==1.16.19
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cors==1.0.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.9.4
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.2
oauthlib==3.1.0
olefile==0.46
pandas==1.4.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.31
systemd-python==234
texttable==1.6.2
tldextract==3.1.2
toml==0.10.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==2.0.2
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
